feat: 전체 계좌 조회 요청 & 계좌 실명 조회 요청 (오픈뱅킹 연동 부분) 구현 및 테스트 (#39, SCRUM-131)

+++ ssok-account-service/src/main/java/kr/ssok/accountservice/entity/enums/BankCode.java
@@ -38,8 +38,8 @@ public enum BankCode {
         }
     }
 
-    public static BankCode fromIdx(Long idx) {
-        BankCode bankCode = IDX_MAP.get(idx.intValue());
+    public static BankCode fromIdx(int idx) {
+        BankCode bankCode = IDX_MAP.get(idx);
         if (bankCode == null) {
             throw new IllegalArgumentException("Invalid BankCode idx: " + idx);
         }
